+++ packages/apollo-server-lambda/src/ApolloServer.ts
@@ -190,7 +190,8 @@ export class ApolloServer extends ApolloServerBase {
       }
 
       if (this.playgroundOptions && event.httpMethod === 'GET') {
-        const acceptHeader = event.headers['Accept'] || event.headers['accept'];
+        const acceptHeader = typeof event.headers === 'object' &&
+          (event.headers['Accept'] || event.headers['accept']);
         if (acceptHeader && acceptHeader.includes('text/html')) {
           const path =
             event.path ||